Moved writing of linear system of #ifdef NOX_DEBUG ... #endif
There is no reason I can think of to not allow writing of the linear
system in an optimized build.  It is often painful to have to recompile
and run in a debug build if you want to save the linear system.  So I
moved the code that saves the linear system out of the ifdef's for a
debug build.  I also put all of the logic inside the if test if writing
is requested, so the only additional overhead is querying the parameter
list.