Fix a dependency issue with stk
If STK was enabled but SEACAS was not, the libstk_io_util.a
library was not built but was required by
stk_util/unit_test_support and caused a build failure at
link time.  Now stk_util/unit_test_support is conditionally
enbled based on whether or not SEACAS is enabled.