don't assume base of chaco file is 1
When Zoltan_Generate_Files writes out chaco-like graph
files using the application's queries, the base ID depends
on what the application uses.  Then when reading the file
in, the chaco reader used the ids in the .graph file for
the neighbor IDs and the range 1..nvtxs for the vertex IDs.

The change is proc 0 figures out the base ID and broadcasts
it to the other procs, and the zdrive mesh build code uses
the base to create the vertex IDs.