Changing script in unit test mode not to have to find git or svn to pass
One of the unit tests failed on the machine p90n03.pbm.ihost.com last night
because it did not find SVN.  This caused the unit test to fail that looks
that the currect SVN command is called.  This change makes the script not have
to find git or svn when in unit testing mode (the right thing to do).

Build/Test Cases Summary
Enabled Packages: TrilinosFramework
Enabled all Forward Packages
0) MPI_DEBUG => passed: passed=86,notpassed=0 (1.15 min)
1) SERIAL_RELEASE => passed: passed=86,notpassed=0 (1.03 min)