Fix issue with configure files not pointing to the correct package configure
There was an issue with subpackages and generating the TrilinosConfig.cmake file
which was causing the configure files from subpackages to be listed as though
they were in the root directory. The issue was that the <subpackage>_BINARY_DIR
variable was not being set in the correct scope for the export system to read it
when generating the TrilinosConfig.cmake file and so it was being treated as
empty.