Fix Issues with dependencies on Thyra
A number of packages which previously depended on Thyra, in part or in whole,
were having issues when the optional parts they depended on were not being
enabled. Mostly there were just tests that couldn't build or link due to missing
dependencies. I have protected the tests that I've found with ifs to ensure that
they are not built when the parts of Thyra that are needed are not enabled.
There may still be a few issues left though.

Build/Test Cases Summary
Enabled Packages: TrilinosFramework, MOOCHO, NOX, Rythmos
Enabled all Forward Packages
0) MPI_DEBUG => passed: passed=306,notpassed=0 (24.98 min)
1) SERIAL_RELEASE => passed: passed=308,notpassed=0 (10.80 min)