Update Moertel to use the new seacas exodus and nemesis subpackages
Changed Moertel to use just seacas exodus and nemesis. Also did s little bit of
clean up of the cmake files.

Added ifpack as a required dependency to moertel since it requires features of
ML that are only enabled when Ifpack is enabled. This was not previously exposed
because optional dependencies are enabled by default.