Fixing failing test Teuchos_TestForException_throwTest_nostacktrace
The test TestForException_throwTest_nostacktrace started failing because I
modified the source code which changed the file and line number of the throw.
However, in my local build with stack tracing enabled, I did not run the full
Teuchos test suite which whould have shown the error.  Instead I just ran the
checkin-test.py script which turns off stack tracing since not every machine
has BinUtils and we want uniform builds across machines with checkin-test.py
builds.  However, the test Teuchos_TestForException_throwTest_nostacktrace was
in an if statement to only be defined when Teuchos_ENABLE_STACKTRACE=ON.
Therefore the checkin-test.py script did not run the test and allowed it to
fail.  That is how this failing test was allowed to be pushed.

To avoid this in the future, I have moved the test
Teuchos_TestForException_throwTest_nostacktrace out of the if statement and
now it always get runs.  It should always produce the right file name and line
number because that is gotten from the C++ pre-processor and is independent of
the compiler options.

Build/Test Cases Summary
Enabled Packages: Teuchos
Disabled Packages: PyTrilinos,TriKota,Optika,Sundance,Stokhos
0) MPI_DEBUG => passed: passed=75,notpassed=0 (0.63 min)
1) SERIAL_RELEASE => passed: passed=72,notpassed=0 (0.48 min)