Disable matlab on nightly build machines
The changes to SEACAS to implement subpackages have allowed it to more correctly
state its dependencies for certain parts of it. This has the side effect of
trying to enable the matlab tpl since a couple subpackages have a required
dependency on the matlab tpl. Since we don't have matlab on any of our testing
machines this is a requirement we cannot currently meet so I am disabling the
tpl for the time being which will prevent the subpackages from being enabled.