Added option --skip-case-no-email and avoid summary of skipped build in final email
With the new handling of PS, SS, and EX code, there will be more cases
where builds will be skipped because no packages are enabled (I will
add better documentation for this in a minute).  Setting the option
--skip-case-no-email will avoid worthless notification emails going
out for skipped builds.  However, the default behavior is maintained.

I also changed the contents of the final summary email to not give the
long summary for buid/test cases where no packages are enabled.  This
now behaves like build/test cases that are explicitly skipped
(e.g. through --without-default-builds).

I added a good unit test for this behavior.