Fixed handling of dg/dx multi-vector orthog poly as an Epetra_Operator.
Things like dg/dx, dg/dxdot, and df/dp or dg/dp when p represents the
input polynomial chaos coefficients need to be represented as
Epetra_Operator's.  These use Stokhos::MatrixFreeOperator as the
operator implementation, which requires each multi-vector coefficient to
appear as an operator.  Instead of manually doing this conversion, I
added an object to store the underlying multi-vector orthog poly, but
appear as an operator orthog poly.  This ensures the coefficient map and
underlying block multi-vector are set and consistent if an underlying
ModelEvaluator tries to use them.  Before they were not set, causing
errors.