Fixing logic so that non-clean working dir fails with --abort-gracefully-if-no-updates and sends error email.
Last week someone pushed changes to a Dependencies.cmake file file but did not
commit the updated *.xml dependency files.  This caused a CI server that moves
commits from SSG to another machine to fail because a push could not occur
because the working directory was not clean.  However, the logic related to
--abort-gracefully-if-no-updates was bad and resulted in no email going out.
Therefore, we did not know that the sync had stopped and it resulted in a lot
of wasted time and effort.

With this commit I added a unit tests that showed the bad behavior and then
changed the code to fix the test.  There are no defects, just missing tests
right?

Build/Test Cases Summary
Enabled Packages: TrilinosFramework
Disabled Packages: PyTrilinos,TriKota,Optika,Sundance,Stokhos
Enabled all Forward Packages
0) MPI_DEBUG => passed: passed=89,notpassed=0 (1.48 min)
1) SERIAL_RELEASE => passed: passed=89,notpassed=0 (1.04 min)
2) MPI_DEBUG_SS => Skipped configure, build, test due to no enabled packages! => Does not affect push readiness! (-1.00 min)
3) SERIAL_RELEASE_SS => Skipped configure, build, test due to no enabled packages! => Does not affect push readiness! (-1.00 min)