Fix an issue with cpack tarball package naming
There were extra decorations being added to the name of the tarball that was
added for the windows installer. These changes make sense for the windows
installer because they make it clear if it was built with mpi or not. I have
moved that naming to be just for the windows build for now.