Adding an MPI_RELEASE build for only Trilinos packages
This build is needed to support the new MPI_RELEASE build for CASL VRI add-on
packages.  We need this build so that if some Trilinos package fails in CASL
VRI testing (sending results to the hidden CASL CDash dashboard), it will also
fail in the native Trilinos build, sending the results to the main Trilinos
CDash dashboard so Trilinos developers can see what is happening and fix the
problem.