Finished complete draft
This version of the document is complete with no more todos.  I read
and edited the entire document and I ran spell check.  This version
should be appropritate to send to the other authors.

There is a lot of duplication in this document and it could likely be
shortended by about 1/4 by moving things around but at least the light
at the end of the tunnel is starting to shine.