Fix NOEXESUFFIX for Windows
Windows builds require the suffix for executables to exist, and be ".exe" so the
option NOEXESUFFIX to PACKAGE_ADD_EXECUTABLE needs to be ignored on Windows
machines. This was causing several build errors in Mesquite, which happens to be
the only package that uses the option and also builds on Windows.

Build/Test Cases Summary
Enabled Packages: TrilinosFramework
Enabled all Packages
0) MPI_DEBUG => passed: passed=1299,notpassed=0 (241.95 min)
1) SERIAL_RELEASE => passed: passed=1312,notpassed=0 (131.65 min)