Put in test to avoid missing CDash XML dependencies file
The CDash XML dependencies file that gets created as part of nightly testing
and CI testing does not get created as part of the 'make dashboard' target.
This was causing a submit failure.  I just added an EXISTS if test to avoid
doing the submit if the file does not exist.