Removing TrilinosCTestDriverCore => TribitsCTestDriverCore token name change
I removed the token name change TrilinosCTestDriverCore =>
TribitsCTestDriverCore because it renamed all of the platform specific driver
script includes (e.g. TrilinosCTestDriverCore.gcc.cmake).  This broke all of
the automated testing.

Will pushed a commit to fix the problem.  This change just ensures that if
someone runs this script again, it will not cause the problem again.