Update Kallikrates build driver script
A few tests in Teuchos and Epetra rely on libraries from other directories. For
those libraries to be properly linked at runtime in the shared build case the
path to them must be on the users %PATH% variable. They are now added. This is
an inelegant solution to this problem though.

Build/Test Cases Summary
Enabled Packages: TrilinosFramework, Teuchos
Enabled all Forward Packages
0) MPI_DEBUG => passed: passed=1235,notpassed=0 (208.09 min)
1) SERIAL_RELEASE => passed: passed=1202,notpassed=0 (117.50 min)
Other local commits for this build/test group: 130916f