Update release tarball creation scripts
Updated a few things on the release tarball scripts to fix a few paths for
trilinos-test instead of trilinos-test2 and make the scripts used to verify the
tarball a bit more like the one used to generate the tarball. One major change
here is that these scripts now build shared so that we can test pytrilinos.