Changed the logic of ${PROJECT_NAME}_UNENABLE_ENABLED_PACKAGES=ON to not touch disabled packages
The logic of ${PROJECT_NAME}_UNENABLE_ENABLED_PACKAGES=ON had to be changed so
that it would not set to empty '' in the cache package enables that were set
hard off.  This was messing up the behavior of the VERA builds in the 'make
dashboard' target where the package disable logic was breaking because setting
the package enables to empty '' in the cache was disabling the logic that was
trying to disable the packages by calling SET( ... CACHE ...) which leaves the
value alone if it is already in the cache.

The change in behavior is the right thing to do and is consistent with the
name of the variable ${PROJECT_NAME}_UNENABLE_ENABLED_PACKAGES=ON.

Build/Test Cases Summary
Enabled Packages: TrilinosFramework, Teuchos
Disabled Packages: PyTrilinos,TriKota,Optika,Sundance,Stokhos
0) MPI_DEBUG => passed: passed=168,notpassed=0 (1.38 min)
1) SERIAL_RELEASE => passed: passed=166,notpassed=0 (1.12 min)