Changes basic names from 'Trilinos' to 'TriBITS'.
A lot more work in the text needs to be done to generalize this from
'Trilinos' to 'TriBITS' but it should not be that bad.  I think that comparing
to the Trilinos 1.0 lifecycle model is still a valid thing to do because what
else are you going to compare to in the CSE community?