Moved default Trilinos git rep locations so they will be read in time.
I had to move the Trilinos git repo location defaults to the
TrilinosCTestDriverCore.cmake and out of the CTestConfig.cmake file because
the file CTestConfig.cmake does not get read until too late.  In my previous
commit, I had these vars in CTestConfig.cmake in an effort to consolidate this
info but it did not work.

Now the system will create the initial clones correctly.