Install the generated version file for all of Trilinos
The Trilinos_version.h file which is generated at configure time was not being
installed for all packages. A few packages(such as teuchos) were adding this
file to their list of headers, which did cause it to be installed in many cases,
but not all. This file should be installed without having to declare it a header
in individual packages so imediately after it is created it is registered for
installation. I have left the packages that do add this file to their list of
headers alone as that does have other consequences in certain build systems
other than make.

Build/Test Cases Summary
Enabled Packages: TrilinosFramework
Enabled all Forward Packages
0) MPI_DEBUG => passed: passed=92,notpassed=0 (5.19 min)
1) SERIAL_RELEASE => passed: passed=92,notpassed=0 (2.66 min)
Other local commits for this build/test group: f520ea5