Fix issues with exclude list for tarball packaging
The introduction of a "." to the paths of files caused the exclude list to not
work since the files to exclude are regular expressions that describe a path and
not paths themselves. The variables that can have a "." in them are resolved to
their actual path before being used in the exclude list now. Also fixed a bug
where PACKAGE_NAME was not being set to the correct package in
TRIBITS_EXCLUDE_FILES causing the wrong files to be excluded.

Package directories also changed to include the "package" part of the path which
made the subset distribution logic that added package to the path redundant so
it has been removed.

cleaned up the autotools exclusion macro some so that it doesn't call the actual
exclude macro so often.

None of the above affected the branch so this has not caused any files that
shouldn't be released to accidentally be put in a tarball.

Build/Test Cases Summary
Enabled Packages: TrilinosFramework
Enabled all Packages
0) MPI_DEBUG => passed: passed=1325,notpassed=0 (164.78 min)
1) SERIAL_RELEASE => passed: passed=1332,notpassed=0 (96.80 min)
Other local commits for this build/test group: 0577bc7