Change the way download-cmake.py finds the host name
This change is necessary for windows as the logic to find a the host name on
windows was to hard code it to the empty string. This change uses the socket
module to get the fully qualified domain name of the current machine. This
greatly simplified the logic since we don't have to call different shell
commands to find the hostname on linux or OS X machines.

I tested this manually on linux mac and windows. In every case the proper http
proxy was set based on the machine name.