Edited the new paragraphs, moved the discussion of AztecOO to Below to grandfathering section
I moved the discussion of the possibility of refactoring AztecOO into
Belos into the "grandfathering" section since I think it goes better
with the discussion of the legacy change algorithm.  I put in the
links between the sections to tie these together.

I also put in some links to the side-note discussion of
quick-and-dirty unit testing.