Merge Fmwk::MeshObj and stk::mesh::Entity, Fmwk::FieldBucket and stk::mesh::Bucket
Issue: We began the migration of Fmwk MeshObj with an implementation that
had a stk entity pointer underneath the MeshObj that could be used to manage
the state that stk entities can manage while leaving the rest to MeshObj.
This proved to be a performance killer since many operations required an
extra dereference (the entity pointer) compared to before, causing lots of
additional cache misses even for simple operations.

The solution is similar to what we did with Relation; we will use the preprocessor
to add in the Fmwk pieces of the class API when the class is being compiled

This change also forced us to merge FieldBucket into stk::mesh::Bucket.

(From Sierra)