fixed an empty row bug ... but perhaps created a new problem??
Basically, there are some moments where having an empty row causes
problems in ML. However, I can't remember what those cases are.
To avoid this, we have generally avoided empty rows in the matrix-matrix
multiply. This is done by putting a 0 in one of the columns. The
problem is determining an appropriate global column index. The
basic idea has been to take a global column index from the rightmost
matrix in a multiply, but if the rightmost matrix is completely empty
I changed the code so that it leaves an empty matrix in the result
of the multiply. Previously, a global column index of 0 was used
if the rightmost matrix was completely empty. This created problems
if processor 0 has no rows as there is no global column id of 0.